test: split template tests and refactor out duplicate code
Almost everything in templateContext was duplicated across the two previous templates test files
